# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TodoList API",
    "version": "1.0.0",
    "description": "개인 할일 관리 애플리케이션 REST API입니다. 사용자 인증(JWT), 할일(Todo) CRUD, 카테고리(Category) CRUD 기능을 제공합니다.\n\n## 인증 방식\n모든 보호된 엔드포인트는 `Authorization: Bearer <token>` 헤더를 요구합니다. 토큰은 `/api/auth/login`을 통해 발급받습니다.\n\n## 비즈니스 규칙\n- 사용자는 자신의 데이터(할일, 카테고리)에만 접근 가능합니다.\n- 동일 사용자 내 카테고리명 중복은 허용되지 않습니다.\n- 할일 상태는 `PENDING`(미완료) 또는 `COMPLETED`(완료)만 존재합니다.\n- 카테고리 삭제 시 해당 카테고리의 할일은 삭제되지 않으며 `categoryId`가 `null`로 변경됩니다.\n- 사용자 탈퇴 시 모든 관련 데이터(카테고리, 할일)가 즉시 영구 삭제됩니다.",
    "contact": {
      "email": "realbackok@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "로컬 개발 서버"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 액세스 토큰. `/api/auth/login` 응답의 `token` 필드 값을 사용합니다."
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "description": "할일 분류 카테고리",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "카테고리 고유 식별자",
            "example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "소유 사용자 ID",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "카테고리명 (동일 사용자 내 중복 불가, 최대 100자)",
            "example": "업무"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "카테고리 생성 일시 (ISO 8601)",
            "example": "2026-04-28T09:00:00.000Z"
          }
        },
        "required": ["id", "userId", "name", "createdAt"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "표준 에러 응답 형식",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "기계가 처리 가능한 에러 코드",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "description": "사람이 읽을 수 있는 에러 메시지",
                "example": "제목은 필수 항목입니다."
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
      "MessageResponse": {
        "type": "object",
        "description": "단순 메시지 응답",
        "properties": {
          "message": {
            "type": "string",
            "description": "처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": ["message"]
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "인증 토큰이 없거나 유효하지 않습니다.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "유효한 인증 토큰이 필요합니다."
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "해당 리소스에 대한 접근 권한이 없습니다.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "이 리소스에 접근할 권한이 없습니다."
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "서버 내부 오류가 발생했습니다.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/account": {
      "delete": {
        "tags": ["Auth"],
        "operationId": "deleteAccount",
        "summary": "회원탈퇴",
        "description": "계정 및 모든 관련 데이터를 영구 삭제합니다. (FR-AUTH-04)\n\n**경고:** 이 작업은 되돌릴 수 없습니다. (BR-AUTH-04)\n- 사용자 계정이 삭제됩니다.\n- 해당 사용자의 모든 카테고리가 즉시 영구 삭제됩니다.\n- 해당 사용자의 모든 할일이 즉시 영구 삭제됩니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "회원탈퇴 성공. 모든 관련 데이터가 삭제되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "계정이 성공적으로 삭제되었습니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/password": {
      "patch": {
        "tags": ["Auth"],
        "operationId": "changePassword",
        "summary": "비밀번호 변경",
        "description": "현재 비밀번호를 확인한 후 새 비밀번호로 변경합니다. (FR-AUTH-05)\n\n**유효성 검사:**\n- 현재 비밀번호가 일치해야 합니다.\n- 새 비밀번호는 최소 8자 이상이어야 합니다.\n- 새 비밀번호와 현재 비밀번호가 달라야 합니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "description": "현재 비밀번호",
                    "example": "password123!"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8,
                    "description": "새 비밀번호 (최소 8자, 현재 비밀번호와 달라야 함)",
                    "example": "newpassword456!"
                  }
                },
                "required": ["currentPassword", "newPassword"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "비밀번호 변경 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "비밀번호가 성공적으로 변경되었습니다."
                }
              }
            }
          },
          "400": {
            "description": "새 비밀번호 유효성 오류 (8자 미만, 현재 비밀번호와 동일)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "tooShort": {
                    "summary": "비밀번호 8자 미만",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "새 비밀번호는 최소 8자 이상이어야 합니다."
                      }
                    }
                  },
                  "sameAsCurrent": {
                    "summary": "현재 비밀번호와 동일",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "새 비밀번호는 현재 비밀번호와 달라야 합니다."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "현재 비밀번호가 일치하지 않거나 인증 토큰이 유효하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "wrongPassword": {
                    "summary": "현재 비밀번호 불일치",
                    "value": {
                      "error": {
                        "code": "INVALID_CREDENTIALS",
                        "message": "현재 비밀번호가 올바르지 않습니다."
                      }
                    }
                  },
                  "unauthorized": {
                    "summary": "인증 토큰 없음/만료",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "유효한 인증 토큰이 필요합니다."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": ["Category"],
        "operationId": "getCategories",
        "summary": "카테고리 목록 조회",
        "description": "로그인한 사용자의 카테고리 목록을 전체 조회합니다. (FR-CAT-02)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "카테고리 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  },
                  "required": ["categories"]
                },
                "example": {
                  "categories": [
                    {
                      "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                      "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "업무",
                      "createdAt": "2026-04-28T09:00:00.000Z"
                    },
                    {
                      "id": "d4e5f6a7-b8c9-0123-defg-234567890123",
                      "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "일상",
                      "createdAt": "2026-04-28T09:05:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Category"],
        "operationId": "createCategory",
        "summary": "카테고리 생성",
        "description": "새 카테고리를 생성합니다. (FR-CAT-01)\n\n**비즈니스 규칙:**\n- 동일 사용자 내에서 카테고리명은 중복될 수 없습니다 (BR-CAT-01).\n- 카테고리명은 최대 100자이며 공백만으로 구성될 수 없습니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "카테고리명 (필수, 최대 100자, 공백만으로 구성 불가, 사용자 내 중복 불가)",
                    "example": "업무"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "카테고리 생성 성공. 생성된 카테고리 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    }
                  },
                  "required": ["category"]
                },
                "example": {
                  "category": {
                    "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                    "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "업무",
                    "createdAt": "2026-04-28T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 오류 (이름 누락 또는 공백)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "카테고리 이름은 필수 항목입니다."
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "description": "동일 사용자 내에 이미 같은 이름의 카테고리가 존재합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "DUPLICATE_NAME",
                    "message": "이미 사용 중인 카테고리 이름입니다."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "tags": ["Category"],
        "operationId": "updateCategory",
        "summary": "카테고리 수정",
        "description": "카테고리명을 수정합니다. (FR-CAT-03)\n\n**비즈니스 규칙:**\n- 본인 소유의 카테고리만 수정할 수 있습니다 (BR-AUTH-03).\n- 변경하려는 이름이 같은 사용자의 다른 카테고리와 중복되면 수정이 실패합니다 (BR-CAT-01).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "수정할 카테고리의 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "수정할 카테고리명 (필수, 최대 100자, 공백만으로 구성 불가, 사용자 내 중복 불가)",
                    "example": "개인 업무"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "카테고리 수정 성공. 수정된 카테고리 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    }
                  },
                  "required": ["category"]
                },
                "example": {
                  "category": {
                    "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                    "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "개인 업무",
                    "createdAt": "2026-04-28T09:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 유효성 오류 (이름 누락 또는 공백)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "카테고리 이름은 필수 항목입니다."
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "해당 ID의 카테고리를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 카테고리를 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "409": {
            "description": "변경하려는 이름의 카테고리가 이미 존재합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "DUPLICATE_NAME",
                    "message": "이미 사용 중인 카테고리 이름입니다."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Category"],
        "operationId": "deleteCategory",
        "summary": "카테고리 삭제",
        "description": "카테고리를 영구 삭제합니다. (FR-CAT-04)\n\n**비즈니스 규칙 (BR-CAT-02):**\n- 본인 소유의 카테고리만 삭제할 수 있습니다 (BR-AUTH-03).\n- 해당 카테고리에 속한 할일들은 삭제되지 않으며, `categoryId`가 `null`로 자동 변경됩니다.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "삭제할 카테고리의 UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
          }
        ],
        "responses": {
          "200": {
            "description": "카테고리 삭제 성공. 연결된 할일의 categoryId는 null로 변경됩니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "카테고리가 성공적으로 삭제되었습니다."
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "해당 ID의 카테고리를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "해당 카테고리를 찾을 수 없습니다."
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  }
}